Prepare package organization for PrometheusAgent
