+++ pkg/prometheus/collector.go
@@ -47,23 +47,23 @@ var (
 	)
 )
 
-type prometheusCollector struct {
+type Collector struct {
 	stores []cache.Store
 }
 
-func newPrometheusCollectorForStores(s ...cache.Store) *prometheusCollector {
-	return &prometheusCollector{stores: s}
+func NewCollectorForStores(s ...cache.Store) *Collector {
+	return &Collector{stores: s}
 }
 
 // Describe implements the prometheus.Collector interface.
-func (c *prometheusCollector) Describe(ch chan<- *prometheus.Desc) {
+func (c *Collector) Describe(ch chan<- *prometheus.Desc) {
 	ch <- descPrometheusSpecReplicas
 	ch <- descPrometheusEnforcedSampleLimit
 	ch <- descPrometheusSpecShards
 }
 
 // Collect implements the prometheus.Collector interface.
-func (c *prometheusCollector) Collect(ch chan<- prometheus.Metric) {
+func (c *Collector) Collect(ch chan<- prometheus.Metric) {
 	for _, s := range c.stores {
 		for _, p := range s.List() {
 			c.collectPrometheus(ch, p.(*v1.Prometheus))
@@ -71,8 +71,8 @@ func (c *prometheusCollector) Collect(ch chan<- prometheus.Metric) {
 	}
 }
 
-func (c *prometheusCollector) collectPrometheus(ch chan<- prometheus.Metric, p *v1.Prometheus) {
-	replicas := float64(minReplicas)
+func (c *Collector) collectPrometheus(ch chan<- prometheus.Metric, p *v1.Prometheus) {
+	replicas := float64(MinReplicas)
 	if p.Spec.Replicas != nil {
 		replicas = float64(*p.Spec.Replicas)
 	}
